Save fixed masureemnts to core data
