+++ AirCasting/CoreData/AirCasting.xcdatamodeld/AirCasting.xcdatamodel/contents
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="17709" systemVersion="20D80" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
+<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="17709" systemVersion="20D64" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
     <entity name="Measurement" representedClassName=".Measurement" syncable="YES" codeGenerationType="class">
         <attribute name="latitude" optional="YES" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
         <attribute name="longitude" optional="YES" attributeType="Double" defaultValueString="0.0" usesScalarValueType="YES"/>
@@ -33,7 +33,7 @@
     </entity>
     <entity name="Session" representedClassName=".Session" syncable="YES" codeGenerationType="class">
         <attribute name="contribute" optional="YES" attributeType="Boolean" defaultValueString="NO" usesScalarValueType="YES"/>
-        <attribute name="deviceId" attributeType="String"/>
+        <attribute name="deviceId" optional="YES" attributeType="String"/>
         <attribute name="deviceType" optional="YES" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
         <attribute name="endTime" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
         <attribute name="followedAt" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
@@ -56,6 +56,6 @@
         <element name="Measurement" positionX="-63" positionY="-18" width="128" height="118"/>
         <element name="MeasurementStream" positionX="-63.7001953125" positionY="128.3192749023438" width="128" height="253"/>
         <element name="SensorThreshold" positionX="388.8433837890625" positionY="56.29141235351562" width="128" height="133"/>
-        <element name="Session" positionX="146.2360229492188" positionY="-4.297317504882812" width="165.6036376953125" height="328"/>
+        <element name="Session" positionX="146.2360229492188" positionY="-4.297317504882812" width="165.6036376953125" height="314"/>
     </elements>
 </model>
