+++ AirCasting/Dashboard/DashboardView.swift
@@ -6,12 +6,13 @@
 //
 
 import SwiftUI
+import CoreData
 
 struct DashboardView: View {
     
     @FetchRequest<Session>(sortDescriptors: [NSSortDescriptor(key: "startTime",
                                                               ascending: false)]) var sessions
-
+    
     var body: some View {
         VStack {
             sectionPicker
